+++ modules/AWSPowerShell/Cmdlets/ApplicationDiscoveryService/Basic/Get-ADSConfiguration-Cmdlet.cs
@@ -46,8 +46,8 @@ namespace Amazon.PowerShell.Cmdlets.ADS
     /// such as host name, operating system, number of network cards, etc.
     /// </para><para>
     /// For a complete list of outputs for each asset type, see <a href="https://docs.aws.amazon.com/application-discovery/latest/userguide/discovery-api-queries.html#DescribeConfigurations">Using
-    /// the DescribeConfigurations Action</a> in the <i>AWS Application Discovery Service
-    /// User Guide</i>.
+    /// the DescribeConfigurations Action</a> in the <i>Amazon Web Services Application Discovery
+    /// Service User Guide</i>.
     /// </para></note>
     /// </summary>
     [Cmdlet("Get", "ADSConfiguration")]