+++ modules/AWSPowerShell/Cmdlets/ApplicationDiscoveryService/Basic/Start-ADSExportTask-Cmdlet.cs
@@ -39,8 +39,9 @@ namespace Amazon.PowerShell.Cmdlets.ADS
     /// agent data is limited to five concurrently running exports. 
     /// </para><para>
     ///  If you do not include an <code>agentIds</code> filter, summary data is exported that
-    /// includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery
-    /// Agents. Export of summary data is limited to two exports per day. 
+    /// includes both Amazon Web Services Agentless Discovery Connector data and summary data
+    /// from Amazon Web Services Discovery Agents. Export of summary data is limited to two
+    /// exports per day. 
     /// </para>
     /// </summary>
     [Cmdlet("Start", "ADSExportTask", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]