+++ modules/AWSPowerShell/Cmdlets/WorkSpacesWeb/Basic/Update-WSWUserSetting-Cmdlet.cs
@@ -51,6 +51,17 @@ namespace Amazon.PowerShell.Cmdlets.WSW
         public Amazon.WorkSpacesWeb.EnabledType CopyAllowed { get; set; }
         #endregion
         
+        #region Parameter DisconnectTimeoutInMinute
+        /// <summary>
+        /// <para>
+        /// <para>The amount of time that a streaming session remains active after users disconnect.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("DisconnectTimeoutInMinutes")]
+        public System.Int32? DisconnectTimeoutInMinute { get; set; }
+        #endregion
+        
         #region Parameter DownloadAllowed
         /// <summary>
         /// <para>
@@ -61,6 +72,18 @@ namespace Amazon.PowerShell.Cmdlets.WSW
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
         [AWSConstantClassSource("Amazon.WorkSpacesWeb.EnabledType")]
         public Amazon.WorkSpacesWeb.EnabledType DownloadAllowed { get; set; }
+        #endregion
+        
+        #region Parameter IdleDisconnectTimeoutInMinute
+        /// <summary>
+        /// <para>
+        /// <para>The amount of time that users can be idle (inactive) before they are disconnected
+        /// from their streaming session and the disconnect timeout interval begins.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("IdleDisconnectTimeoutInMinutes")]
+        public System.Int32? IdleDisconnectTimeoutInMinute { get; set; }
         #endregion
         
         #region Parameter PasteAllowed
@@ -191,7 +214,9 @@ namespace Amazon.PowerShell.Cmdlets.WSW
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
             context.ClientToken = this.ClientToken;
             context.CopyAllowed = this.CopyAllowed;
+            context.DisconnectTimeoutInMinute = this.DisconnectTimeoutInMinute;
             context.DownloadAllowed = this.DownloadAllowed;
+            context.IdleDisconnectTimeoutInMinute = this.IdleDisconnectTimeoutInMinute;
             context.PasteAllowed = this.PasteAllowed;
             context.PrintAllowed = this.PrintAllowed;
             context.UploadAllowed = this.UploadAllowed;
@@ -226,9 +251,17 @@ namespace Amazon.PowerShell.Cmdlets.WSW
             {
                 request.CopyAllowed = cmdletContext.CopyAllowed;
             }
+            if (cmdletContext.DisconnectTimeoutInMinute != null)
+            {
+                request.DisconnectTimeoutInMinutes = cmdletContext.DisconnectTimeoutInMinute.Value;
+            }
             if (cmdletContext.DownloadAllowed != null)
             {
                 request.DownloadAllowed = cmdletContext.DownloadAllowed;
+            }
+            if (cmdletContext.IdleDisconnectTimeoutInMinute != null)
+            {
+                request.IdleDisconnectTimeoutInMinutes = cmdletContext.IdleDisconnectTimeoutInMinute.Value;
             }
             if (cmdletContext.PasteAllowed != null)
             {
@@ -309,7 +342,9 @@ namespace Amazon.PowerShell.Cmdlets.WSW
         {
             public System.String ClientToken { get; set; }
             public Amazon.WorkSpacesWeb.EnabledType CopyAllowed { get; set; }
+            public System.Int32? DisconnectTimeoutInMinute { get; set; }
             public Amazon.WorkSpacesWeb.EnabledType DownloadAllowed { get; set; }
+            public System.Int32? IdleDisconnectTimeoutInMinute { get; set; }
             public Amazon.WorkSpacesWeb.EnabledType PasteAllowed { get; set; }
             public Amazon.WorkSpacesWeb.EnabledType PrintAllowed { get; set; }
             public Amazon.WorkSpacesWeb.EnabledType UploadAllowed { get; set; }